Module hotshot_web_server::config

Configuration for the webserver

Constants

DEFAULT_WEB_SERVER_DA_PORT

the default port on which to serve Data availability functionality

DEFAULT_WEB_SERVER_PORT

the default port on which to run the web server

DEFAULT_WEB_SERVER_VIEW_SYNC_PORT

the default port on which to serve View Sync functionality

MAX_TXNS

How many transactions to keep in memory

MAX_VIEWS

How many views to keep in memory

TX_BATCH_SIZE

How many transactions to return at once

Functions

get_da_certificate_route

get latest certificate

get_latest_proposal_route

get latest qc

get_latest_view_sync_certificate_route

get latest view sync proposal

get_proposal_route

get proposal

get_transactions_route

get transactions

get_upgrade_route

get upgrade route

get_vid_certificate_route

get vid certificate

get_vid_disperse_route

get vid dispersal

get_vid_vote_route

get vid vote route

get_view_sync_certificate_route

get view sync proposal

get_view_sync_vote_route

get view sync vote

get_vote_route

get vote

post_da_certificate_route

post data availability certificate

post_proposal_route

post proposal

post_staketable_route

post stake table

post_transactions_route

post transactions

post_upgrade_route

post upgrade route

post_vid_certificate_route

post vid certificate

post_vid_disperse_route

post vid dispersal

post_vid_vote_route

post vid vote route

post_view_sync_certificate_route

post view sync proposal

post_view_sync_vote_route

post view sync vote

post_vote_route

post vote